Basic feature (commit by Huy)

+++ ManagementInternet/Models/Entities/ComputerType.cs
@@ -21,7 +21,7 @@ namespace ManagementInternet.Models.Entities
         [StringLength(255)]
         public string Name { get; set; }
 
-        public decimal Price { get; set; }
+        public short Price { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Computer> Computers { get; set; }
