+++ ManagementInternet/Models/Entities/Order.cs
@@ -15,7 +15,7 @@ namespace ManagementInternet.Models.Entities
             PlayTimeManagements = new HashSet<PlayTimeManagement>();
         }
 
-        public short Id { get; set; }
+        public int Id { get; set; }
 
         public short IdOfComputer { get; set; }
